+++ app/src/main/java/com/example/youtubeapi/data/local/AppDatabase.kt
@@ -5,7 +5,7 @@ import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import com.example.youtubeapi.data.local.dao.VideoDao
-import com.example.youtubeapi.data.model.VideoEntity
+import com.example.youtubeapi.data.model.entity.VideoEntity
 
 @Database(entities = [VideoEntity::class], version = 1)
 abstract class AppDatabase: RoomDatabase() {
