+++ app/src/main/java/com/example/youtubeapi/data/remote/RemoteDataSource.kt
@@ -1,7 +1,7 @@
 package com.example.youtubeapi.data.remote
 
 import com.example.youtubeapi.BuildConfig
-import com.example.youtubeapi.data.model.VideoResponse
+import com.example.youtubeapi.data.model.dto.VideoResponse
 import retrofit2.http.GET
 import retrofit2.http.Query
