+++ app/src/main/java/com/example/youtubeapi/presentation/MainActivity.kt
@@ -6,7 +6,7 @@ import com.example.youtubeapi.HomeFragment_TAG
 import com.example.youtubeapi.MyVideoFragment_TAG
 import com.example.youtubeapi.R
 import com.example.youtubeapi.SearchFragment_TAG
-import com.example.youtubeapi.adapter.ViewPagerAdapter
+import com.example.youtubeapi.presentation.adapter.ViewPagerAdapter
 import com.example.youtubeapi.databinding.ActivityMainBinding
 import com.google.android.material.tabs.TabLayoutMediator
